use-cases/agent-factory-with-subagents/agents/pwa_mobile_agent/requirements.txt
# PWA Mobile Agent dependencies
pydantic-ai>=0.0.7
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
aiofiles>=23.0.0  # File operations
httpx>=0.24.0     # HTTP requests
jinja2>=3.1.0     # Template generation